Split up Graphene binding code
