+++ src/a_tuin/constants.py
@@ -1,0 +1,7 @@
+__copyright__ = 'Copyright(c) Gordon Elliott 2017'
+
+""" 
+"""
+
+# identify db session in app server request context
+SESSION = 'session'
