+++ src/glod/api/account_node.py
@@ -5,7 +5,12 @@ __copyright__ = 'Copyright(c) Gordon Elliott 2017'
 
 import graphene
 
-from a_tuin.api import node_connection_field, get_update_mutation, get_create_mutation
+from a_tuin.api import (
+    node_class,
+    node_connection_field,
+    get_update_mutation,
+    get_create_mutation
+)
 from glod.api.account_leaf import AccountLeaf, account_fields
 from glod.api.fund_node import FundLeaf
 from glod.db.account import Account, AccountQuery
@@ -17,10 +22,11 @@ account_node_fields['funds'] = graphene.Field(
     graphene.List(FundLeaf, description='Funds for this account.')
 )
 
-AccountNode, accounts_connection_field = node_connection_field(
+AccountNode = node_class(Account.__name__, AccountLeaf, account_node_fields)
+
+accounts_connection_field = node_connection_field(
     AccountQuery,
-    AccountLeaf,
-    account_node_fields,
+    AccountNode,
     description='List of all bank accounts'
 )
 CreateAccountLeaf = get_create_mutation(Account, account_fields, AccountLeaf)
