+++ src/glod/api/subject_node.py
@@ -3,15 +3,21 @@ __copyright__ = 'Copyright(c) Gordon Elliott 2017'
 """ 
 """
 
-from a_tuin.api import node_connection_field, get_update_mutation, get_create_mutation
+from a_tuin.api import (
+    node_class,
+    node_connection_field,
+    get_update_mutation,
+    get_create_mutation
+)
 from glod.api.subject_leaf import SubjectLeaf, subject_fields
 from glod.db.subject import Subject, SubjectQuery
 
 
-SubjectNode, subjects_connection_field = node_connection_field(
+SubjectNode = node_class(Subject.__name__, SubjectLeaf, subject_fields)
+
+subjects_connection_field = node_connection_field(
     SubjectQuery,
-    SubjectLeaf,
-    subject_fields,
+    SubjectNode,
     description="List of all subjects"
 )
 CreateSubjectLeaf = get_create_mutation(Subject, subject_fields, SubjectLeaf)
